@@ -2,7 +2,7 @@ package ddl.itlab.ncku.twmonument
 
 import android.content.Intent
 import android.os.Bundle
-import android.support.v7.app.AppCompatActivity
+import androidx.appcompat.app.AppCompatActivity
 import android.view.Gravity
 import android.widget.TextView
 import android.widget.Toast
