@@ -4,13 +4,9 @@ import android.annotation.SuppressLint
 import android.content.Context
 import android.content.Intent
 import android.net.Uri
-import android.os.Build
 import android.os.Bundle
-import android.support.annotation.RequiresApi
-import android.support.v7.app.AppCompatActivity
+import androidx.appcompat.app.AppCompatActivity
 import android.view.MenuItem
-import android.webkit.WebChromeClient
-import android.webkit.WebResourceRequest
 import android.webkit.WebView
 import android.webkit.WebViewClient
 
@@ -38,7 +34,7 @@ class WebActivity : AppCompatActivity() {
         val url = intent.getStringExtra(INTENT_URL)
 
         // setup title
-        this.title = Uri.parse(url).host.toString()
+        this.title = Uri.parse(url)?.host.toString()
 
         // setup webview
         findViewById<WebView>(R.id.webView1).let {
